feat: Update user profile and file upload

+++ src/controllers/videoController.js
@@ -1,33 +1,45 @@
 import Video, { formatHashtags } from "../models/Video";
-
+import User from "../models/user";
 export const home = async (req, res) => {
     const videos = await Video.find({}).sort({createdAt: "desc"});
     return res.render("home", { pageTitle: `Home` , videos });
 }
 export const watch = async (req, res) => {
     const { id } = req.params;
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("owner");
     if(!video) {
         return res.render("404", { pageTitle: "Video not found." });
     }
-    return res.render("watch", {pageTitle: `Watching`, video });
+    return res.render("watch", { pageTitle: `Watching`, video });
 };
 
 export const getEdit = async (req, res) => {
     const { id } = req.params;
+    const { 
+        user: { _id },
+    } = req.session;
     const video = await Video.findById(id);
     if(!video) {
         return res.render("404", { pageTitle: "Video not found." });
+    }
+    if(String(video.owner) !== _id) {
+        return res.status(403).redirect("/");
     }
     return res.render("edit", { pageTitle: `Edit: ${video.title}`, video });
 };
 
 export const postEdit = async (req, res) => {
     const { id } = req.params;
+    const { 
+        user: { _id },
+    } = req.session;
     const {title, description, hashtags} = req.body;
     const video = await Video.exists({ _id: id });
     if(!video) {
         return res.render("404", { pageTitle: "Video not found." });
+    }
+    if(String(video.owner) !== _id) {
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id, {
         title, description, 
@@ -42,18 +54,28 @@ export const getUpload = (req, res) => {
 };
 
 export const postUpload = async (req, res) => {
+    const { 
+        user: {_id}
+    } = req.session;
+
+    const { path:fileUrl } = req.file; // ES6
     const {title, description, hashtags } = req.body;
     try {
-        await Video.create({
+        const newVideo = await Video.create({
             title,
             description,
+            fileUrl,
+            owner: _id,
             hashtags: Video.formatHashtags(hashtags),
         });    
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
         return res.redirect("/");   
     } catch (error) {
         console.log(error);
         return res.render("upload", {  
-            pageTitle: "Upload Vidoe", 
+            pageTitle: "Upload Video", 
             errorMessage: error._message,
         });
     }
@@ -61,6 +83,16 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async(req, res) => {
     const { id } = req.params;
+    const { 
+        user: { _id },
+    } = req.session;
+    const video = await Video.findById(id);
+    if(!video) {
+        return res.render("404", { pageTitle: "Video not found." });
+    }
+    if(String(video.owner) !== _id) {
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
 }
